aseanL-main/requirements.txt
openai-whisper
speechrecognition
pyaudio
soundfile
